api-producer/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
celery==5.5.2
pika==1.3.2
psycopg2-binary==2.9.10
redis==6.0.0